+++ CHANGELOG.md
@@ -17,6 +17,8 @@
   array of text fragments with support for fallback glyphs
 * Configuration option 'font.on_invalid_glyph' for use together with the new
   method for creating text fragments with fallback glyphs
+* Configuration option 'font.fallback' which is used by the default
+  implementation of 'font.on_invalid_glyph'
 * [HexaPDF::Document::Layout#text_fragments] for creating text fragments with
   support for fallback glyphs via 'font.on_invalid_glyph'
 * [HexaPDF::Content::CanvasComposer] for using high-level layout functionality
@@ -26,7 +28,6 @@
   for determining whether a font is bold
 * [HexaPDF::Font::TrueTypeWrapper#italic?] and
   [HexaPDF::Font::Type1Wrapper#italic?] for determining whether a font is italic
-
 
 ### Changed